bma_send_whatsapp/requirements_binary.txt
requests==2.32.5
mysql-connector-python==9.5.0

